Tests/LibWeb/Text/expected/HTML/iframe-successfully-loads-in-shadow-root.txt

Received a message: 'Hello from iframe in the shadow root of the just inserted div!'
Was it from the shadow root iframe? true
